Fixing OS issues.
git-svn-id: https://rabbit-eclipse.googlecode.com/svn/trunk@35 03b96cd2-0a55-11df-9859-c12d74d802c5

+++ Rabbit UI/src/rabbit/ui/internal/pages/CommandPageLabelProvider.java
@@ -48,7 +48,7 @@ public class CommandPageLabelProvider extends LabelProvider implements ITableLab
 			case 1:
 				return cmd.getDescription();
 			case 2:
-				return page.getValue(cmd) + "";
+				return (int) page.getValue(cmd) + "";
 			default:
 				return null;
 			}
